backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
anthropic==0.40.0
google-api-python-client==2.150.0
google-auth==2.35.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
python-dotenv==1.0.0
python-multipart==0.0.12
httpx==0.27.0
supabase==2.9.1
